added support for keeping apple double ghost files

+++ lib/unmacer.rb
@@ -7,6 +7,7 @@ class Unmacer
                 :keep_fsevents,
                 :keep_trashes,
                 :keep_macosx,
+                :keep_apple_double,
                 :keep_apple_double_orphans
 
   SPOTLIGHT = '.Spotlight-V100'
@@ -21,6 +22,7 @@ class Unmacer
     self.keep_fsevents             = false
     self.keep_trashes              = false
     self.keep_macosx               = false
+    self.keep_apple_double         = false
     self.keep_apple_double_orphans = false
   end
 
@@ -62,7 +64,7 @@ private
   def unmac_folder(dirname)
     delete_macosx(dirname) unless keep_macosx
     delete_ds_store(dirname)
-    delete_apple_double(dirname)
+    delete_apple_double(dirname) unless keep_apple_double
   end
 
   def delete(parent, file_or_directory)
